Move the Python value-conversion layer from pydantic-monty to monty-proto
The Python <-> MontyObject conversion layer (convert/dataclass/exception
conversions) now lives in monty-proto behind a new off-by-default `python`
cargo feature, exposed as a flat `monty_proto::python` module. This removes
the rlib backwards-coupling where monty-cpython linked all of pydantic-monty
(and transitively monty-pool/tokio) just for those three modules:

- monty-cpython depends on `monty-proto/python` instead of pydantic-monty
- pydantic-monty is back to a pure cdylib and keeps only its Python-facing
  exception class hierarchy in exceptions.rs
- the `python` feature never enables pyo3's `extension-module`, so the top
  crate still decides how libpython links; pure-Rust consumers stay pyo3-free
- `cfg!(Py_3_XX)` checks became runtime `py.version_info()` checks since
  monty-proto has no pyo3-build-config build script (and the runtime
  interpreter version is the correct thing to test anyway)
