+++ crates/monty-cpython/Cargo.toml
@@ -23,12 +23,8 @@ path = "src/lib.rs"
 ahash = { workspace = true }
 clap = { workspace = true }
 monty = { workspace = true }
-monty-proto = { workspace = true }
-# `pydantic-monty`'s lib (`_monty`) is linked as an rlib purely for its
-# Python ↔ MontyObject value-conversion layer (`convert`/`dataclass`/`exceptions`).
-# This works because that crate does not enable pyo3's `extension-module` feature
-# in Cargo, so libpython is linked and we can embed CPython here via `auto-initialize`.
-pydantic-monty = { path = "../monty-python" }
+# `python` enables monty-proto's Python ↔ MontyObject value-conversion layer
+monty-proto = { workspace = true, features = ["python"] }
 pyo3 = { version = "0.28", features = ["auto-initialize"] }
 # Blocking WebSocket transport for the `websocket` subcommand: dial a relay
 # (or a parent-as-server), with rustls for `wss://`. Matches the pool client.
