+++ crates/monty-proto/Cargo.toml
@@ -19,6 +19,12 @@ name = "monty_proto"
 monty = { workspace = true }
 num-bigint = { workspace = true }
 prost = { workspace = true }
+# Optional: powers the `python` feature's Python ↔ MontyObject conversion layer
+# (`src/python/`), shared by `pydantic-monty` (the extension module) and
+# `monty-cpython` (the embedded-CPython child). `extension-module` is never
+# enabled here — the top crate decides how libpython is linked (maturin enables
+# it when building wheels).
+pyo3 = { version = "0.28", optional = true, features = ["num-bigint"] }
 # generation-only dependencies (see src/bin/generate.rs); exact pins so the
 # checked-in generated code is reproducible across contributors and CI
 prost-build = { version = "=0.14.4", optional = true }
@@ -28,6 +34,10 @@ protox = { version = "=0.9.1", optional = true }
 # Enables the `generate-proto` binary that regenerates src/generated from the
 # .proto schema. Not used by normal builds — the generated code is checked in.
 generate = ["dep:prost-build", "dep:protox"]
+# Enables the `python` module: PyO3-based conversions between Python objects
+# and `MontyObject`/`MontyException`. Off by default so pure-Rust consumers
+# (monty-pool, monty-runtime) never build or link against libpython.
+python = ["dep:pyo3"]
 
 [[bin]]
 name = "generate-proto"
