+++ crates/monty-proto/src/lib.rs
@@ -3,6 +3,10 @@
 mod convert;
 mod frame;
 mod generated;
+// Python ↔ MontyObject value conversion; opt-in because it links pyo3, which
+// pure-Rust consumers of the wire protocol must never pay for.
+#[cfg(feature = "python")]
+pub mod python;
 mod requirement;
 mod wire;
