+++ crates/monty-python/src/pool.rs
@@ -35,6 +35,7 @@ use std::{
 
 use ::monty::{ExcType, ExtFunctionResult, MontyException, MontyObject};
 use monty_pool::{Checkout, MountSpec, MountSpecMode, Pool, PoolConfig, PoolError, ReplConfig, ResumeValue, TurnEvent};
+use monty_proto::python::{DcRegistry, exc_py_to_monty, monty_to_py, py_to_monty_value};
 use pyo3::{
     exceptions::{PyRuntimeError, PyTimeoutError, PyTypeError, PyValueError},
     prelude::*,
@@ -46,9 +47,7 @@ use tokio::task::{JoinSet, spawn_blocking};
 use crate::{
     async_dispatch::{dispatch_function_call, join_error_to_py, spawn_coroutine_task, wait_for_futures},
     build::{extract_repl_inputs, extract_source_code, extract_type_check_stubs},
-    convert::{monty_to_py, py_to_monty_value},
-    dataclass::DcRegistry,
-    exceptions::{MontyCrashedError, MontyError, MontyTypingError, exc_py_to_monty},
+    exceptions::{MontyCrashedError, MontyError, MontyTypingError},
     external::{CallResult, ExternalLookup, dispatch_method_call},
     get_not_handled,
     limits::extract_limits,
